+++ UVtools.GUI/Forms/FrmSettings.Designer.cs
@@ -143,10 +143,12 @@
             this.label39 = new System.Windows.Forms.Label();
             this.nmCrosshairLineMargin = new System.Windows.Forms.NumericUpDown();
             this.groupBox9 = new System.Windows.Forms.GroupBox();
+            this.cbShowLayerTooltipsOverlay = new System.Windows.Forms.CheckBox();
             this.cbLayerAutoRotateBestView = new System.Windows.Forms.CheckBox();
             this.cbLayerZoomToFit = new System.Windows.Forms.CheckBox();
             this.tabPage3 = new System.Windows.Forms.TabPage();
             this.tabPage4 = new System.Windows.Forms.TabPage();
+            this.cbCloseEditOnApply = new System.Windows.Forms.CheckBox();
             this.groupBox10 = new System.Windows.Forms.GroupBox();
             this.btnPixelEditorDrainHoleColor = new System.Windows.Forms.Button();
             this.btnPixelEditorDrainHoleHLColor = new System.Windows.Forms.Button();
@@ -173,8 +175,6 @@
             this.cbLayerRepairLayersIslands = new System.Windows.Forms.CheckBox();
             this.panel1 = new System.Windows.Forms.Panel();
             this.toolTip = new System.Windows.Forms.ToolTip(this.components);
-            this.cbCloseEditOnApply = new System.Windows.Forms.CheckBox();
-            this.cbHidePLTooltips = new System.Windows.Forms.CheckBox();
             this.groupBox3.SuspendLayout();
             ((System.ComponentModel.ISupportInitialize)(this.nmResinTrapBinaryThreshold)).BeginInit();
             ((System.ComponentModel.ISupportInitialize)(this.nmResinTrapMaximumPixelBrightnessToDrain)).BeginInit();
@@ -1654,7 +1654,7 @@
             // 
             // groupBox9
             // 
-            this.groupBox9.Controls.Add(this.cbHidePLTooltips);
+            this.groupBox9.Controls.Add(this.cbShowLayerTooltipsOverlay);
             this.groupBox9.Controls.Add(this.cbLayerAutoRotateBestView);
             this.groupBox9.Controls.Add(this.cbLayerZoomToFit);
             this.groupBox9.Location = new System.Drawing.Point(5, 483);
@@ -1663,6 +1663,17 @@
             this.groupBox9.TabIndex = 56;
             this.groupBox9.TabStop = false;
             this.groupBox9.Text = "Miscellaneous";
+            // 
+            // cbShowLayerTooltipsOverlay
+            // 
+            this.cbShowLayerTooltipsOverlay.AutoSize = true;
+            this.cbShowLayerTooltipsOverlay.Location = new System.Drawing.Point(393, 27);
+            this.cbShowLayerTooltipsOverlay.Name = "cbShowLayerTooltipsOverlay";
+            this.cbShowLayerTooltipsOverlay.Size = new System.Drawing.Size(206, 22);
+            this.cbShowLayerTooltipsOverlay.TabIndex = 4;
+            this.cbShowLayerTooltipsOverlay.Text = "Show preview layer tooltips";
+            this.toolTip.SetToolTip(this.cbShowLayerTooltipsOverlay, "Show or hide layer preview tooltips for select, ROI, and pixel edit modes.");
+            this.cbShowLayerTooltipsOverlay.UseVisualStyleBackColor = true;
             // 
             // cbLayerAutoRotateBestView
             // 
@@ -1678,7 +1689,7 @@
             // cbLayerZoomToFit
             // 
             this.cbLayerZoomToFit.AutoSize = true;
-            this.cbLayerZoomToFit.Location = new System.Drawing.Point(217, 27);
+            this.cbLayerZoomToFit.Location = new System.Drawing.Point(208, 27);
             this.cbLayerZoomToFit.Name = "cbLayerZoomToFit";
             this.cbLayerZoomToFit.Size = new System.Drawing.Size(152, 22);
             this.cbLayerZoomToFit.TabIndex = 3;
@@ -2070,28 +2081,6 @@
             this.toolTip.ToolTipIcon = System.Windows.Forms.ToolTipIcon.Info;
             this.toolTip.ToolTipTitle = "Information";
             // 
-            // cbCloseEditOnApply
-            // 
-            this.cbCloseEditOnApply.AutoSize = true;
-            this.cbCloseEditOnApply.Location = new System.Drawing.Point(11, 213);
-            this.cbCloseEditOnApply.Name = "cbCloseEditOnApply";
-            this.cbCloseEditOnApply.Size = new System.Drawing.Size(322, 22);
-            this.cbCloseEditOnApply.TabIndex = 25;
-            this.cbCloseEditOnApply.Text = "Close pixel editor after operations are applied";
-            this.toolTip.SetToolTip(this.cbCloseEditOnApply, "Automatically close pixel editor after applying changes");
-            this.cbCloseEditOnApply.UseVisualStyleBackColor = true;
-            // 
-            // cbHidePLTooltips
-            // 
-            this.cbHidePLTooltips.AutoSize = true;
-            this.cbHidePLTooltips.Location = new System.Drawing.Point(404, 27);
-            this.cbHidePLTooltips.Name = "cbHidePLTooltips";
-            this.cbHidePLTooltips.Size = new System.Drawing.Size(198, 22);
-            this.cbHidePLTooltips.TabIndex = 4;
-            this.cbHidePLTooltips.Text = "Hide preview layer tooltips";
-            this.toolTip.SetToolTip(this.cbHidePLTooltips, "Show or hide layer preview tooltips for select, ROI, and pixel edit modes.");
-            this.cbHidePLTooltips.UseVisualStyleBackColor = true;
-            // 
             // FrmSettings
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(9F, 18F);
@@ -2304,6 +2293,6 @@
         private System.Windows.Forms.Button btnPixelEditorAddPixelColor;
         private System.Windows.Forms.Button btnPixelEditorAddPixelHLColor;
         private System.Windows.Forms.CheckBox cbCloseEditOnApply;
-        private System.Windows.Forms.CheckBox cbHidePLTooltips;
+        private System.Windows.Forms.CheckBox cbShowLayerTooltipsOverlay;
     }
 }